Renamed the account-metadata-service webpp and refactored its packages
