+++ components/account-metadata-service/src/main/java/org/wso2/openbanking/consumerdatastandards/account/metadata/service/exceptions/AccountMetadataException.java
@@ -1,0 +1,44 @@
+/**
+ * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.openbanking.consumerdatastandards.account.metadata.service.exceptions;
+
+/**
+ * Exception for account metadata operations.
+ */
+public class AccountMetadataException extends Exception {
+
+    /**
+     * Constructs a new exception with the specified message.
+     *
+     * @param message the error message
+     */
+    public AccountMetadataException(String message) {
+        super(message);
+    }
+
+    /**
+     * Constructs a new exception with the specified message and cause.
+     *
+     * @param message the error message
+     * @param cause the cause of the exception
+     */
+    public AccountMetadataException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
