--- components/account-metadata-webapp/src/main/java/org/wso2/openbanking/consumerdatastandards/endpoints/disclosure/options/model/ApiResponse.java
@@ -1,50 +1,0 @@
-/**
- * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.openbanking.consumerdatastandards.endpoints.disclosure.options.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-
-/**
- * Generic API response wrapper for JSON responses.
- */
-@ApiModel(description = "API response object")
-public class ApiResponse {
-
-    @ApiModelProperty(
-            value = "Response message",
-            example = "Disclosure options updated successfully"
-    )
-    @JsonProperty("message")
-    private String message;
-
-    public ApiResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-}
